Add the release information for CI jobs
The release information is now contained in the releases directory.
This file is used by the CI jobs when moving a docker image
from the staging repo to the release repo.

The file currently contains "dummy" information. Once an image is
generated, a change to this file with the correct reference hash,
and version number will be pushed.

Issue-ID: RICAPP-106

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Icad4f4f8f9b25bbe0d5b0827d504563de58a1fef