Update patch set 3
Patch Set 3:

This change introduces the releases file.  Once the CI jobs are in and active, we'll need to force a remerge to have it build the image.  Once the image is built, we can change the release file to trigger the promotion of the image to release.  

Please review, verify and merge. Thanks.

Patch-set: 3
Reviewer: Gerrit User 1000371 <1000371@84e4e190-e623-4c37-9eda-58b1937c1b53>