Update patch set 1
Patch Set 1: Code-Review-1

There needs to be a tox.ini file in the root directory.  I don't fully understand tox, nor why when run in the docs directory it looks in the parent, but it does.  

The _static directory in ./docs is missing.   This needs to have a logo.png file in it.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>
Label: Code-Review=-1